Update README.md
This repository includes the work that was needed to be done as the project work done in order to get the intern certificate from CodeClause.

This repository includes a Login Page for sign-in and sign-up, which will then take you to the homepage which includes info about me and my earlier projects as the computer science student of Indian Institute of Information Technology, Kottayam during the 1st and 2nd year of my college. Further it includes the hyperlink to all the 4 projects which are needed to be done in order to complete the training period at CodeClause. The 4 projects that are done are: VERSE : Which is the name to the web app which is Live Language translator NUCLEO : a basic text editor MUSEMIND : a expense tracker FIREART : a music streaming application

The hyperlinks will take you directly to the Landing page of the respective project page which is further created using HTML, CSS, JavaScript, ReactJS, VanillaJS, etc. The landing page are made to increase and show the richness of the stuff that was created. And i assume that they definitely helped me to showcase the required skill and skillset. Some of the landing page and the project webpage includes the DarkMode that will help the use to toggle between the modes as per therir suitability